Construct and transmit frames out of the serial port
This patch adds a function to sric-if.c that constructs the next frame
to go out of the serial port.  This is then used when the port's fd is
ready for writing.

At the moment, it doesn't wait for a response back from the target of
the frame before transmitting the next from the queue.