Return from read_frames after processing one, for good reason:
In the model that was developing, tokens only advance around the bus
when we've exhausted messages to process. This means an unpleasent
race situation can develop where we instruct a device to pass the
token onwards during enumeration, but that doesn't immediately occur,
and the following frame gets processed with the token in an
unexpected position.

So only process one frame at a time, advance token after each.