Add beginnings of how enum FSM will recognize each device:
The plan in my mind: every 50ms (or thereabouts) poll the gateway
to see whether it's received the token. If it has, stop. If it
hasn't, there's an unenumerated device out there, so we perform
enum steps.

Should any message loss occur, the timeout will instigate the
same device getting enumerated again. Precisely how this pans
out remains to be seen.