Send token to new addr, not broadcast, due to races:
If we have two devices on the bus, and they're not executing
simultaniously, we can reach the unpleasent situation where both
devices process tok_advance; the first sends the token on to the
second; which then checks whether it has the token, and advances
it.

So one tok_advance message would result in two devices advancing
the token.