Ignore SIGPIPE signals
SIGPIPE is sent to processes when one tries to write to a pipe that
is not open at the other end.  The default action when this happens is
for the process to be terminated.  This is obviously not what we want.

This patch sets the SIGPIPE signal to be ignored.  When the signal is
ignored, the write will return an error as normal, resulting in a sane
client disconnection -- rather than sricd termination!