Provoke any devices holding the token on start-up to release their token
The old version of libsric that was shipped with the servo board
doesn't let tokens that it's holding through when it's sent RESET.
This can lead to enumeration failing much.  Unfortunately, because
enumeration fails, there's no way we could reliably update the
firmware in the servo boards, as there'd be no guarantee that we were
_only_ talking to a servo board!

This means we have to get creative with the tools that the shipped
servo firmware offers us.  Luckily, we can get it to let go of its
token using TOK_ADVANCE.

This patch causes sricd to send two TOK_ADVANCE frames out at the
beginning of enumeration.  This causes the boards to give up their
tokens, so that we can get on with the rest of our lives.

It's not possible to send out more TOK_ADVANCE frames, because the
token regeneration in the power board starts to get involved.  (I've
tested this: with token regeneration, if you take too long, then
enumeration doesn't work.)  So, this only sends out two.  Since the
bus has actually been extremely reliable, and there's no motor
activity occuring during enumeration, we can just about get away with
this.

If it weren't a week from the competition, I would engineer a more
advanced updating scenario, in which we added a flag to sricd to
indicate whether it should send out many or few TOK_ADVANCE frames.
It'd then update the power board's firmware, restart sricd with
different flags, update the servo board, and get on with its life.
Would probably involve adding some kind of firmware versioning
mechanism to the gateway commands, as we can send them without
enumeration.  Anyway, I digress.