Add an option to ignore the first SIGTERM received
When the USB key is yanked from our robots, all the processes relating
to it are sent a SIGTERM, then a SIGKILL a little later.  We want
sricd to hang around a bit so that we can do things like get motors to
stop etc.  This patch allows sricd to optionally ignore the first
SIGTERM, which is exactly what we need for this application.